web/requirements.txt
falcon==2.0.0
falcon-cors==1.1.7
falcon-multipart==0.2.0
gevent==1.4.0
gunicorn==19.9.0
pika==1.1.0
python-magic==0.4.15
